Experiment 3: more exploration
Using longer pre-learning random play time and eps anneal to collect more diverse training data.

Previously, the reward curve always shows a small bump and dip before the bigger increase to the global maximum. The hypothesis is that the first dip was because that the first round of learning was based on insufficient exploration, which wasn't enough to push it to the global maximum. The dip helped collect more exploration data, which causes the second increase hit global maximum.

After changing the config here, the initial small bump disappeared, which seems to validate the hypothesis.